refactor(studio): extract buildFleet into engine/fleet.ts
Continues the engine.ts Phase 2 split started by f4ac17c (types, freight).
buildFleet is a pure synchronous function (no IO, no diagnostics, no
Snowflake) so it lifts cleanly into engine/fleet.ts and becomes trivially
unit-testable. engine.ts shrinks from 1118 -> 1058 LOC.

Verified with: tsc -p tsconfig.server.json --noEmit (clean).
